UnixTime: use int64, not uint64
This is more consistent with what Go does, and since we haven't released
this yet, now is the best time to make the change!

+++ cmd/aggregator/main.go
@@ -51,7 +51,7 @@ func (s *server) Publish(ctx context.Context, point *api.PublishRequest) (*api.P
 		return nil, fmt.Errorf("no metadata")
 	}
 
-	ts := time.Unix(int64(point.UnixTime), 0)
+	ts := time.Unix(point.UnixTime, 0)
 
 	tags := map[string]string{}
